.5.1: Add BaseGame.cpp
