+++ native/cocos/platform/java/modules/SystemWindow.h
@@ -40,7 +40,6 @@ public:
     void      copyTextToClipboard(const std::string& text) override;
     uintptr_t getWindowHandler() const override;
     Size      getViewSize() const override;
-    void      setViewSize(uint32_t w, uint32_t h) override {}
 };
 
 } // namespace cc
